template styling updates
