+++ src/components/Template/registry.tsx
